# PM benchmarks & optimizations