… benchmarks/pattern_matching/array.rb
… benchmarks/pattern_matching/hash.rb
… benchmarks/pattern_matching/mixed.rb
… benchmarks/pattern_matching/struct.rb