### What is the purpose of this pull request?

Benchmark and optimize pattern matching.

----

This PR adds some pattern matching optimizations making transpiled code ~2x faster than the MRI's one (2.7.0) for array patterns and about the same for hash patterns.

```
# benchmarks/pattern_matching/array.rb

## Before
Comparison:
          transpiled:   1374101.7 i/s
            baseline:    958536.1 i/s - 1.43x  slower

# After
Comparison:
          transpiled:   1775531.6 i/s 
            baseline:   1000877.7 i/s - 1.77x  slower

# benchmarks/pattern_matching/mixed.rb

## Before
Comparison:
            baseline:    316929.1 i/s
          transpiled:    272592.9 i/s - 1.16x  slower

# After
Comparison:
          transpiled:    320370.3 i/s 
            baseline:    303103.5 i/s - 1.06x  slower

# benchmarks/pattern_matching/hash.rb

# Before
Comparison:
  baseline (first match):   3298961.5 i/s
transpiled (first match):   2561716.8 i/s - 1.29x  slower

Comparison:
  baseline (second match):    921737.8 i/s
transpiled (second match):    753798.9 i/s - 1.22x  slower

# After
Comparison:
  baseline (first match):   3437976.3 i/s 
transpiled (first match):   3021784.2 i/s - 1.14x  slower

Comparison:
  baseline (second match):    943056.7 i/s 
transpiled (second match):    925693.6 i/s - 1.02x  slower

# benchmarks/pattern_matching/struct.rb

# Before
Comparison:
            baseline:    990408.9 i/s
          transpiled:    865842.3 i/s - 1.14x  slower

# After
Comparison:
            baseline:   1069539.1 i/s 
          transpiled:   1029457.8 i/s - 1.04x  slower

# benchmarks/pattern_matching/cases/call.rb

# Before
Comparison:
transpiled (last pattern):   1424591.8 i/s 
  baseline (last pattern):    786053.3 i/s - 1.81x  slower

# After
Comparison:
transpiled (last pattern):   1587132.5 i/s 
  baseline (last pattern):    756756.4 i/s - 2.10x  slower
```